refactoring ver 0.2
