+++ app/src/main/java/com/nick_sib/popularlibraries/mvp/presenters/list/IListPresenter.kt
@@ -1,0 +1,11 @@
+package com.nick_sib.popularlibraries.mvp.presenters.list
+
+import com.nick_sib.popularlibraries.mvp.view.IItemView
+import com.nick_sib.popularlibraries.mvp.view.UserItemView
+
+interface IListPresenter<V : IItemView> {
+    var itemClickListener: ((V) -> Unit)?
+    fun bindView(view: V)
+    fun getCount(): Int
+}
+
