+++ app/src/main/java/com/nick_sib/popularlibraries/mvp/presenters/list/IUserListPresenter.kt
@@ -1,0 +1,5 @@
+package com.nick_sib.popularlibraries.mvp.presenters.list
+
+import com.nick_sib.popularlibraries.mvp.view.UserItemView
+
+interface IUserListPresenter : IListPresenter<UserItemView>
