+++ app/src/main/java/com/nick_sib/popularlibraries/mvp/view/IItemView.kt
@@ -1,0 +1,5 @@
+package com.nick_sib.popularlibraries.mvp.view
+
+interface IItemView {
+    var pos: Int
+}
