+++ app/src/main/java/com/nick_sib/popularlibraries/mvp/view/MainView.kt
@@ -1,11 +1,9 @@
-package com.nick_sib.popularlibraries.ui.users
+package com.nick_sib.popularlibraries.mvp.view
 
 import moxy.MvpView
 import moxy.viewstate.strategy.AddToEndSingleStrategy
 import moxy.viewstate.strategy.StateStrategyType
 
+
 @StateStrategyType(AddToEndSingleStrategy::class)
-interface UsersView : MvpView {
-    fun init()
-    fun updateList()
-}
+interface MainView : MvpView
