+++ app/src/main/java/com/nick_sib/popularlibraries/mvp/view/UserItemView.kt
@@ -1,0 +1,6 @@
+package com.nick_sib.popularlibraries.mvp.view
+
+
+interface UserItemView: IItemView {
+    fun setLogin(text: String)
+}
