+++ app/src/main/java/com/nick_sib/popularlibraries/mvp/view/UsersView.kt
@@ -1,9 +1,11 @@
-package com.nick_sib.popularlibraries.ui.main
+package com.nick_sib.popularlibraries.mvp.view
 
 import moxy.MvpView
 import moxy.viewstate.strategy.AddToEndSingleStrategy
 import moxy.viewstate.strategy.StateStrategyType
 
-
 @StateStrategyType(AddToEndSingleStrategy::class)
-interface MainView : MvpView
+interface UsersView : MvpView {
+    fun init()
+    fun updateList()
+}
