--- app/src/main/java/com/nick_sib/popularlibraries/ui/adapter/IItemView.kt
@@ -1,5 +1,0 @@
-package com.nick_sib.popularlibraries.ui.adapter
-
-interface IItemView {
-    var pos: Int
-}
