+++ app/src/main/java/com/nick_sib/popularlibraries/ui/fragments/UsersFragment.kt
@@ -1,4 +1,4 @@
-package com.nick_sib.popularlibraries.ui.users
+package com.nick_sib.popularlibraries.ui.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,7 +6,9 @@ import android.view.View
 import android.view.ViewGroup
 import com.nick_sib.popularlibraries.App
 import com.nick_sib.popularlibraries.R
-import com.nick_sib.popularlibraries.model.GithubUsersRepo
+import com.nick_sib.popularlibraries.mvp.model.GithubUsersRepo
+import com.nick_sib.popularlibraries.mvp.presenters.UsersPresenter
+import com.nick_sib.popularlibraries.mvp.view.UsersView
 import com.nick_sib.popularlibraries.ui.adapter.UsersRVAdapter
 import kotlinx.android.synthetic.main.fragment_users.*
 import moxy.MvpAppCompatFragment
