--- app/src/main/java/com/nick_sib/popularlibraries/ui/users/UserItemView.kt
@@ -1,8 +1,0 @@
-package com.nick_sib.popularlibraries.ui.users
-
-import com.nick_sib.popularlibraries.ui.adapter.IItemView
-
-
-interface UserItemView: IItemView {
-    fun setLogin(text: String)
-}
